Add files via upload
PS4: SpecterDev releases 4.05 Kernel Exploit

BY WOLOLO · DECEMBER 27, 2017



It’s finally here! 2 Months after Team Fail0verflow revealed the technical details of a Kernel Exploit for firmware 4.05, Developer SpecterDev released a functional implementation today.

The release is fresh and I haven’t taken the time to test it yet (plus, I’m on 4.01 right now and will need to update), but there’s no reason at this point to doubt this release. If you’re running a PS4 4.05, please give it a try and report in the comments.


Related: How to get your hands on a PS4 running firmware 4.05
Understand that this release, in its current state, might not be 100% user friendly. But, fingers crossed, this should open up the PS4 scene a little bit more, and hopefully new developments will make this very interesting.

Specifically, SpecterDev states:



In this project you will find a full implementation of the “namedobj” kernel exploit for the PlayStation 4 on 4.05. It will allow you to run arbitrary code as kernel, to allow jailbreaking and kernel-level modifications to the system. This release however, does not contain any code related to defeating anti-piracy mechanisms or running homebrew. This exploit does include a loader that listens for payloads on port 9020 and will execute them upon receival.

Credits

SpecterDev credits the following people for their contribution to the exploit (besides Fail0verflow):

qwertyoruiopz
Flatz
CTurt
Anonymous
Download and install PS4 4.05 Kernel Exploit
You can download the Files on SpecterDev’s github here.

Additionally, The hacker has released a sample payload to enable the Debug settings. you can find it here.

To run the exploit, you need to have a console on firmware 4.05. If you are below 4.05, you can update to 4.05 by finding the right update file online. If you are above 4.05, you cannot downgrade and this exploit will not work for you.

You’ll need to host the exploit files on a local server, and access the index.html file through Webkit on the PS4. Once the exploit runs successfully, the console will be listening for payloads on port 9020. You can send the payloads with command netcat from your computer.

I apologize as my explanation’s very rough for now. More as we dive into it ourselves!

Source: SpecterDev